[추가] 기본로직 추가, yml 설정파일 추가
