+++ cafekiosk/src/main/java/sample/cafekiosk/spring/CafekioskApplication.java
@@ -2,9 +2,11 @@ package sample.cafekiosk.spring;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 import java.util.Scanner;
 
+@EnableJpaAuditing
 @SpringBootApplication
 public class CafekioskApplication {
