+++ cafekiosk/src/main/java/sample/cafekiosk/spring/domain/product/Product.java
@@ -1,0 +1,29 @@
+package sample.cafekiosk.spring.domain.product;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import sample.cafekiosk.spring.domain.BaseEntity;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Entity
+public class Product extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String productNumber;
+
+    @Enumerated(EnumType.STRING)
+    private ProductType type;
+
+    @Enumerated(EnumType.STRING)
+    private ProductSellingStatus sellingStatus;
+
+    private String name;
+
+    private int price;
+}
