+++ cafekiosk/src/main/java/sample/cafekiosk/spring/domain/product/ProductRepository.java
@@ -1,0 +1,19 @@
+package sample.cafekiosk.spring.domain.product;
+
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface ProductRepository extends JpaRepository<Product, Long>{
+
+    /**
+     * select *
+     * from product
+     * where selling_type in ('SELLING,'HOLD')
+     */
+
+    List<Product> findAllBySellingStatusIn(List<ProductSellingStatus> sellingTypes);
+}
