+++ cafekiosk/src/main/java/sample/cafekiosk/spring/domain/product/ProductSellingStatus.java
@@ -1,0 +1,20 @@
+package sample.cafekiosk.spring.domain.product;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+import java.util.List;
+
+@Getter
+@RequiredArgsConstructor
+public enum ProductSellingStatus{
+    SELLING("판매중"),
+    HOLD("판매보류"),
+    STOP_SELLING("판매중지");
+
+    private final String text;
+
+    public static List<ProductSellingStatus> forDisplay() {
+        return List.of(SELLING, HOLD);
+    }
+}
