+++ cafekiosk/src/main/java/sample/cafekiosk/spring/domain/product/ProductType.java
@@ -1,0 +1,14 @@
+package sample.cafekiosk.spring.domain.product;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum ProductType {
+    HANDMADE("제조 음료"),
+    BOTTLING("병음료"),
+    BAKERY("베이커리");
+
+    private final String text;
+    }
