+++ cafekiosk/src/main/resources/data.sql
@@ -1,0 +1,4 @@
+insert into product(product_number, type, selling_status, name, price)
+values ('001', 'HANDMADE', 'SELLING', '아메리카노', 4000),
+       ('002', 'HANDMADE', 'HOLD', '카페라떼', 4500),
+       ('003', 'BAKERY', 'STOP_SELLING', '크루아상', 3500);
