+++ readable/src/main/java/studycafe/asis/StudyCafeApplication.java
@@ -1,0 +1,10 @@
+package studycafe.asis;
+
+public class StudyCafeApplication {
+
+    public static void main(String[] args) {
+        StudyCafePassMachine studyCafePassMachine = new StudyCafePassMachine();
+        studyCafePassMachine.run();
+    }
+
+}
