+++ readable/src/main/java/studycafe/asis/exception/AppException.java
@@ -1,0 +1,9 @@
+package studycafe.asis.exception;
+
+public class AppException extends RuntimeException {
+
+    public AppException(String message) {
+        super(message);
+    }
+
+}
