+++ readable/src/main/java/studycafe/assignment/StudyCafeApplication.java
@@ -1,0 +1,10 @@
+package readable.src.main.java.studycafe.assignment;
+
+public class StudyCafeApplication {
+
+    public static void main(String[] args) {
+        StudyCafePassMachine studyCafePassMachine = new StudyCafePassMachine();
+        studyCafePassMachine.run();
+    }
+
+}
