+++ readable/src/main/java/studycafe/assignment/exception/AppException.java
@@ -1,0 +1,9 @@
+package readable.src.main.java.studycafe.assignment.exception;
+
+public class AppException extends RuntimeException {
+
+    public AppException(String message) {
+        super(message);
+    }
+
+}
