+++ readable/src/main/java/studycafe/assignment/model/StudyCafeLockerPass.java
@@ -1,0 +1,44 @@
+package readable.src.main.java.studycafe.assignment.model;
+
+public class StudyCafeLockerPass {
+
+    private final StudyCafePassType passType;
+    private final int duration;
+    private final int price;
+
+    private StudyCafeLockerPass(StudyCafePassType passType, int duration, int price) {
+        this.passType = passType;
+        this.duration = duration;
+        this.price = price;
+    }
+
+    public static StudyCafeLockerPass of(StudyCafePassType passType, int duration, int price) {
+        return new StudyCafeLockerPass(passType, duration, price);
+    }
+
+    public StudyCafePassType getPassType() {
+        return passType;
+    }
+
+    public int getDuration() {
+        return duration;
+    }
+
+    public int getPrice() {
+        return price;
+    }
+
+    public String display() {
+        if (passType == StudyCafePassType.HOURLY) {
+            return String.format("%s시간권 - %d원", duration, price);
+        }
+        if (passType == StudyCafePassType.WEEKLY) {
+            return String.format("%s주권 - %d원", duration, price);
+        }
+        if (passType == StudyCafePassType.FIXED) {
+            return String.format("%s주권 - %d원", duration, price);
+        }
+        return "";
+    }
+
+}
