--- src/main/java/readable/studycafe/asis/exception/AppException.java
@@ -1,9 +1,0 @@
-package readable.studycafe.asis.exception;
-
-public class AppException extends RuntimeException {
-
-    public AppException(String message) {
-        super(message);
-    }
-
-}
