--- src/main/java/readable/studycafe/asis/io/InputHandler.java
@@ -1,42 +1,0 @@
-package readable.studycafe.asis.io;
-
-
-
-import readable.studycafe.asis.exception.AppException;
-import readable.studycafe.asis.model.StudyCafePass;
-import readable.studycafe.asis.model.StudyCafePassType;
-
-import java.util.List;
-import java.util.Scanner;
-
-public class InputHandler {
-
-    private static final Scanner SCANNER = new Scanner(System.in);
-
-    public StudyCafePassType getPassTypeSelectingUserAction() {
-        String userInput = SCANNER.nextLine();
-
-        if ("1".equals(userInput)) {
-            return StudyCafePassType.HOURLY;
-        }
-        if ("2".equals(userInput)) {
-            return StudyCafePassType.WEEKLY;
-        }
-        if ("3".equals(userInput)) {
-            return StudyCafePassType.FIXED;
-        }
-        throw new AppException("잘못된 입력입니다.");
-    }
-
-    public StudyCafePass getSelectPass(List<StudyCafePass> passes) {
-        String userInput = SCANNER.nextLine();
-        int selectedIndex = Integer.parseInt(userInput) - 1;
-        return passes.get(selectedIndex);
-    }
-
-    public boolean getLockerSelection() {
-        String userInput = SCANNER.nextLine();
-        return "1".equals(userInput);
-    }
-
-}
