--- src/main/java/readable/studycafe/assignment/StudyCafeApplication.java
@@ -1,10 +1,0 @@
-package readable.studycafe.assignment;
-
-public class StudyCafeApplication {
-
-    public static void main(String[] args) {
-        StudyCafePassMachine studyCafePassMachine = new StudyCafePassMachine();
-        studyCafePassMachine.run();
-    }
-
-}
