--- src/main/java/readable/studycafe/assignment/exception/AppException.java
@@ -1,9 +1,0 @@
-package readable.studycafe.assignment.exception;
-
-public class AppException extends RuntimeException {
-
-    public AppException(String message) {
-        super(message);
-    }
-
-}
