--- src/main/java/readable/studycafe/assignment/model/StudyCafePassType.java
@@ -1,15 +1,0 @@
-package readable.studycafe.assignment.model;
-
-public enum StudyCafePassType {
-
-    HOURLY("시간 단위 이용권"),
-    WEEKLY("주 단위 이용권"),
-    FIXED("1인 고정석");
-
-    private final String description;
-
-    StudyCafePassType(String description) {
-        this.description = description;
-    }
-
-}
